requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.4
Pillow>=10.0.0
pytesseract>=0.3.10